requirements.txt
Django~=2.0.6
djangorestframework==3.9.2
Pygments==2.3.1
Markdown==3.1